fix: address PR #284 review nits
- Compute read tool count dynamically in the writes-enabled manifest
  description (was hardcoded 17, would silently lie if read tools change).
- Extract shared `Manifest` types and `truncateDescription` to a new
  `scripts/manifest-utils.ts`. Removes the `as never` cast in
  sync-manifest.ts and the duplicated truncate helper.
- Add a regression test that the description reports an accurate
  read/write split.

The `args`-absent edge case raised by the bot is intentionally not
guarded — the MCP manifest spec requires `mcp_config.args` to be a
string array, and adding defensive null checks for impossible states
is anti-pattern per project guidance.
