+++ scripts/sync-manifest.ts
@@ -15,6 +15,7 @@
 
 import { createToolSchemas } from '../src/tools/tools.js';
 import { buildWriteManifest } from './build-write-manifest.js';
+import { truncateDescription, type Manifest, type ManifestTool } from './manifest-utils.js';
 import { readFileSync, writeFileSync } from 'fs';
 import { join, dirname } from 'path';
 import { fileURLToPath } from 'url';
@@ -23,52 +24,12 @@ const __dirname = dirname(fileURLToPath(import.meta.url));
 const manifestPath = join(__dirname, '../manifest.json');
 const writeManifestPath = join(__dirname, '../manifest.write.json');
 
-interface ManifestTool {
-  name: string;
-  description: string;
-}
-
-interface Manifest {
-  tools: ManifestTool[];
-  [key: string]: unknown;
-}
-
-function truncateDescription(description: string, maxLength: number = 150): string {
-  // Handle empty or whitespace-only descriptions
-  if (!description || !description.trim()) {
-    return 'No description available.';
-  }
-
-  const trimmed = description.trim();
-
-  // Find the first sentence-ending period (followed by space, end of string, or newline)
-  // This avoids splitting on periods in abbreviations like "e.g." or "etc."
-  const sentenceEndMatch = trimmed.match(/^(.+?\.)\s|^(.+?\.)$/);
-  const firstSentence = sentenceEndMatch
-    ? (sentenceEndMatch[1] || sentenceEndMatch[2])
-    : null;
-
-  // If we found a sentence and it fits within maxLength, use it
-  if (firstSentence && firstSentence.length <= maxLength) {
-    return firstSentence;
-  }
-
-  // Otherwise, truncate at maxLength
-  if (trimmed.length <= maxLength) {
-    // Short description without period - add one
-    return trimmed.endsWith('.') ? trimmed : trimmed + '.';
-  }
-
-  // Truncate long descriptions with ellipsis
-  return trimmed.slice(0, maxLength - 3).trimEnd() + '...';
-}
-
 function main() {
   const writeMode = process.argv.includes('--write');
   const manifest: Manifest = JSON.parse(readFileSync(manifestPath, 'utf-8'));
 
   if (writeMode) {
-    const writeManifest = buildWriteManifest(manifest as never) as unknown as Manifest;
+    const writeManifest = buildWriteManifest(manifest);
     writeFileSync(writeManifestPath, JSON.stringify(writeManifest, null, 2) + '\n');
     console.log(
       `✓ Wrote manifest.write.json with ${writeManifest.tools.length} tools (writes-enabled, local-only)`
